added reference binding to interface ref_of (generated functions). This allows to bind referenced pointer-targets by Ref_t, using an optional logical argument "bind". updated makefile and test-codes

+++ test_simulator_driver.f90
@@ -69,7 +69,7 @@ program sim_driver
 
   subroutine callback()
     real*8, pointer :: t
-    integer*8       :: t_val
+    real*8       :: t_val
 
     t => _getValRef( 't', real8 )
     t_val = _getVal( 't' )
